# target/generated-sources/src/main/java/com/manulife/ap/rsf_connect_mongodb/swagger/api/DoctorsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.0).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.manulife.ap.rsf_connect_mongodb.swagger.api;

import com.manulife.ap.rsf_connect_mongodb.swagger.model.Doctor;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;

@Api(value = "doctors", description = "the doctors API")
public interface DoctorsApi {

    @ApiOperation(value = "Create a new doctor", nickname = "createDoctor", notes = "Insert a new doctor into MongoDB", response = Doctor.class, tags={ "doctors", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Doctor created successfully", response = Doctor.class) })
    @RequestMapping(value = "/doctors",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Doctor> createDoctor(@ApiParam(value = "Doctor object to be created" ,required=true )  @Valid @RequestBody Doctor doctor);


    @ApiOperation(value = "Delete doctor", nickname = "deleteDoctor", notes = "Delete a doctor by ID", tags={ "doctors", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "Doctor deleted successfully"),
        @ApiResponse(code = 404, message = "Doctor not found") })
    @RequestMapping(value = "/doctors/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteDoctor(@ApiParam(value = "Doctor ID",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "Get all doctors", nickname = "getAllDoctors", notes = "Retrieve all doctors from MongoDB", response = Doctor.class, responseContainer = "List", tags={ "doctors", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation", response = Doctor.class, responseContainer = "List") })
    @RequestMapping(value = "/doctors",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Doctor>> getAllDoctors();


    @ApiOperation(value = "Get doctor by ID", nickname = "getDoctorById", notes = "Retrieve a specific doctor by ID", response = Doctor.class, tags={ "doctors", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation", response = Doctor.class),
        @ApiResponse(code = 404, message = "Doctor not found") })
    @RequestMapping(value = "/doctors/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Doctor> getDoctorById(@ApiParam(value = "Doctor ID",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "Update doctor", nickname = "updateDoctor", notes = "Update an existing doctor", response = Doctor.class, tags={ "doctors", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Doctor updated successfully", response = Doctor.class),
        @ApiResponse(code = 404, message = "Doctor not found") })
    @RequestMapping(value = "/doctors/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Doctor> updateDoctor(@ApiParam(value = "Doctor ID",required=true) @PathVariable("id") String id,@ApiParam(value = "Updated doctor object" ,required=true )  @Valid @RequestBody Doctor doctor);

}