# target/generated-sources/src/main/java/com/manulife/ap/rsf_connect_mongodb/swagger/model/Doctor.java
package com.manulife.ap.rsf_connect_mongodb.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.io.Serializable;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * Doctor
 */
@Validated

public class Doctor  implements Serializable {
  private static final long serialVersionUID = 1L;

  @JsonProperty("id")
  private String id = null;

  @JsonProperty("name")
  private String name = null;

  @JsonProperty("gender")
  private String gender = null;

  public Doctor id(String id) {
    this.id = id;
    return this;
  }

  /**
   * Unique identifier (MongoDB ObjectId)
   * @return id
  **/
  @ApiModelProperty(value = "Unique identifier (MongoDB ObjectId)")


  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }

  public Doctor name(String name) {
    this.name = name;
    return this;
  }

  /**
   * Doctor's name
   * @return name
  **/
  @ApiModelProperty(required = true, value = "Doctor's name")
  @NotNull


  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public Doctor gender(String gender) {
    this.gender = gender;
    return this;
  }

  /**
   * Doctor's gender
   * @return gender
  **/
  @ApiModelProperty(value = "Doctor's gender")


  public String getGender() {
    return gender;
  }

  public void setGender(String gender) {
    this.gender = gender;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Doctor doctor = (Doctor) o;
    return Objects.equals(this.id, doctor.id) &&
        Objects.equals(this.name, doctor.name) &&
        Objects.equals(this.gender, doctor.gender);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, name, gender);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Doctor {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    gender: ").append(toIndentedString(gender)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}